requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.1
langchain>=0.1.0
pinecone-client>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
flask>=2.3.0
twilio>=8.10.0

